backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
python-docx>=1.1.0
PyPDF2>=3.0.0

